Introduces VirtualMachineFleet and VirtualMachineFleetLive
